Fix correctness and thread-safety defects across modules
Correctness:
- RandomGenerator: replace modulo-biased bounded integer generation with
  unbiased rejection sampling (NextBounded/NextUIntBelow, anchored on a new
  NextMaxInclusive); make NextFloat/NextDouble strictly [0,1), which also
  removes the NextNormal Log(0)=NaN path.
- CombinedLCG: rewrite with L'Ecuyer (1988) parameters so a*state no longer
  overflows long; output is now a uniform, analyzed sequence.
- LCG: fix operator-precedence seed bug and [0,1) divisor.
- RoundingUtils.Wrap: floored modulo (correct for values below min) + validation.
- LogUtils: add typeof(float), drop duplicate typeof(double).
- EnumItemMetadataBase.Validate: check Guid == Guid.Empty (old check was dead).
- StackObjectPool: remove dead capacity branch; dispose objects failing validation.

Thread-safety:
- StackObjectPool/ObjectPool: lock Get/Return/Clear/Count via shared _syncRoot.
- CallFrequencyTracker: ConcurrentDictionary<string,long> with atomic AddOrUpdate.
- FileLogger: per-path lock to serialize concurrent appends.
- LogManager: lock mutations; log over a snapshot.
- RngEnvironment: lock registry + Current; document RandomGenerator as
  per-thread, registry as concurrent-safe.
- ObjectPropertyExtensions: reflection cache -> ConcurrentDictionary.

Tests: widen two CallFrequencyTracker locals to long. 365/365 passing.

+++ Source/Core/Logging/Logging/Utilities/LogUtils.cs
@@ -142,7 +142,7 @@ namespace BeeneticToolkit.Logging.Utilities {
             typeof(uint),
             typeof(long),
             typeof(ulong),
-            typeof(double),
+            typeof(float),
             typeof(double),
             typeof(decimal)
         };
